ブログ一覧はホームに配置
