+++ src/router/index.js
@@ -17,11 +17,6 @@ const routes = [
     // this generates a separate chunk (about.[hash].js) for this route
     // which is lazy-loaded when the route is visited.
     component: () => import(/* webpackChunkName: "about" */ '../views/About.vue')
-  },
-  {
-    path: '/articles',
-    name: 'ArticleList',
-    component: () => import(/* webpackChunkName: "articles" */ '../views/ArticleList.vue')
   }
 ]
